backend/requirements.txt
pinecone-client
typing
openai
llama-index
flask
flask-cors
transformers
torch
python-dotenv
